float trouble

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: dfinnis <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/05 15:06:08 by dfinnis           #+#    #+#             */
-/*   Updated: 2018/12/09 14:31:05 by svaskeli         ###   ########.fr       */
+/*   Updated: 2018/12/09 19:38:53 by svaskeli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -79,7 +79,7 @@ void				ft_print(t_print *all, va_list ap);
 
 void				ft_print_struct(t_print *all); //rm!!!!!!!!!
 
-long double	ft_modulus(long double a, long double b);
+double	ft_modulus(long double a, long double b);
 unsigned int	ft_len(long double n);
 int	ft_len_int(long long n);
 char	*ft_itoa_float(long double n);
